+++ examples/generative/corrdiff/shell_scripts/example_patch_generate.sh
@@ -1,0 +1,1 @@
+RUN_CMD="python -u generate.py res_ckpt_filename=training-state-diffusion.mdlus sampling_method=stochastic num_steps=18 overlap_pix=4 boundary_pix=2 patch_shape_x=64 patch_shape_y=64"